LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/inline-block-002-expected.txt




FAIL Anchor selection can select empty inline-blocks assert_equals: expected 200 but got 150
